refactor(verifier): split aci/verifier.rs into focused submodules
Break the 2372-line src/aci/verifier.rs into a directory module:

- mod.rs       — module wiring, re-exports, default-timeout consts, and the
                 shared hex/time helpers (decode_hex, decode_hex_32, ...)
- simple.rs    — Static/Preverified verifiers
- external.rs  — ExternalProviderVerifier (Python bridge) machinery
- providers.rs — per-provider wrappers + routing verifier
- report.rs    — ACI attestation-report binding validation
- dcap.rs      — native ACI/DCAP verifier + cache + policy
- dstack.rs    — dstack event-log / RTMR / KMS-custody helpers
- tests.rs     — the existing unit tests

Pure relocation: no logic changes. Public paths preserved via re-export;
cross-module helpers marked pub(super). cargo test --all-targets:
220 passed, 0 failed (unchanged from baseline).
